requirements.txt
Django~=5.2
djangorestframework
django-cors-headers
djangorestframework-simplejwt
mysqlclient
Pillow
gunicorn
python-decouple
drf-yasg